CompilationFix: revert use of "nullptr" to a 0
One build version on Travis C.I. (Linux/Cmake/GCC) for a different code
branch did not seem to like the use of nullptr - so I have reverted to the
ubiquitous numeric 0 instead.  This is because it is a C++11 feature and
there is no requirement to use C++11 specified in the CMake project file.

In passing also noted that a few used files ./src/irc/ircglobal.h,
./src/irc/IRCGlobal & ./src/irc/IrcSender are used but not listed in the
qmake ./src/src.pro project file - so they do not show up in the Qt Creator
IDE - and the first contains the version number of the Communi IRC library
that we incorporate and use.  So, I've added it to the project file as
well.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>