CompilationFix: revert use of "nullptr" to a 0
One build version on Travis C.I. (Linux/Cmake/GCC) did not seem to like the
use of nullptr - so I have reverted to the ubiquitous numeric 0 instead.
This is because it is a C++11 feature and there is no requirement to use
C++11 specified in the CMake project file.

In passing also noted that a used file ./src/irc/ircglobal.h is used but
not listed in the qmake ./src/src.pro project file - so it does not show up
in the Qt Creator IDE - and it contains the version number of the
LibIrcClient-Qt library that we incorporate and use.  So, I've added it to
the project file as well.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>