Fix: Ensure that default (-1) area ALWAYS exists
Although we have the provision of a default area (with Id -1) we have not
previously actually made sure it exists after loading (or clearing) prior
to loading in a map - which causes all sorts of error messages when a map
WITHOUT that area is loaded.

This will possible need checking with a planned revision to code to allow
user loading of XML map files in the near future.

Also:
* moved all instance in TMap class of mpHost->postMessage( QString ) to own
  postMessage( QString ) implimentation
* removed unused QDataStream reference argument in:
  * (void) TRoomDB::restoreSingleArea( QDataStream &, int, TArea * )
  * (void) TRoomDB::restoreSingleRoom( QDataStream &, int, TRoom * )
* removed unused Globals in XMLimport.cpp:
  * int maxAreas
  * QMap<int,int> areaMap

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>