BugFix: correct bug that caused removal of doors on stub exits
Also refactored TLuaInterpreter::setDoor(...) & getDoor(...) to current
error handling standards - including checks for validity. The name of the
normal exit directions always were specific but are now explicitly checked
for - this goes in step with the recently added auditing code that removes
doors on exits that do not exist (even as a normal direction "stub" exit).

The setDoor(...) function returns true if a change was made to the room's
doors and false if the command was ineffective.  It now also returns a
secondary error message if a nil is returned to explain what the problem
was...!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>