UserRequest: Allow Default Area to be hidden in mapper area control
As per this message in the Forum:
http://forums.mudlet.org/viewtopic.php?f=13&t=4857&start=20#p24479
now that a recent commit correctly ensures that a TArea for the default
area Id of -1 IS instantiated - they have noticed that it shows up in the
area control widget (a non-editable QComboBox) whereas it had not in the
past. However the original poster has a need to HIDE that area from casual
inspection by the user of their package.  As the "fix" of that recent
commit does affect the application behaviour this commit provides both
script-able and GUI means to select whether to show the Default area in
that widget.

The setting DOES not persist between settings (it would not be easy to save
in a profile specific way without introducing a new setting to the game
save format - which is NOT currently versioned {unlike the Map file format}
and it is not really a map setting).

Also, it cannot be applied until a map display (mapper) is created - so the
option is hidden on the "Profile Preferences" => "Special Options" tab and
the lua command "setDefaultAreaVisible(bool)" will return false until there
is one; after that the checkbox control is shown on the dialog and will
return true from the lua command.  On profile loading/mapper creation the
Default Area WILL be shown in the widget - corresponding to a true value
to the lua command.  Using the lua command will immediately refresh the
list of Areas in that widget if the setting is changed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>