Enhance: make area selection widget centre map on room in middle of area
As discussed in http://forums.mudlet.org/viewtopic.php?f=13&t=4818#p24498
this commit improves the way the centre of the view is selected when the
area selection widget is used to (temporarily) show a view of a different
area:
* A) if the selected area has the same z-coordinate as the current view
in the (2D) mapper then the mean coordinate of all the rooms on THAT level
is calculated and then the room nearest to that is chosen to be at the
centre of the view in the selected area.
* B) if the selected area does not have any rooms on the same level then
the level with the MOST rooms (and the lowest z coordinate if there are
ties) is chosen and the room closest to the mean again selected.
* C) if there are NOT rooms in the area then the view is centred on the
origin.
* D) if (after using the control more than once OR even if NOT) the
original area is selected then the view snaps back to the current player
room (this can be used to recenter on that if the scroll buttons have been
used to offset the view!)

The center of view position is made from code within the 2D mapper - but
there is no need to do the same for the 3D one - instead the area ID and
position is passed up to the TMap class instance which forwards the details
to the 3D mapper.

The bottom line is that (unless it is an empty area) a reasonable choice
of room is ALWAYS displayed in the centre of the map when the area is
changed manually (and the 2 and 3D maps will be centred on the same room).

Code associated with loading maps has been refactored in an attempt to
ensure that the area selection widget is set to the correct value before
the first use by the user - there was four places where the map file
was being loaded (although the one in the Host class constructor seems
unnecessary - and problematic to handle as the Host class is instantiated
before the mapper widget is created!) I *think* I have got it straight now
but some testing in the map down load from server case may need a check to
see that I have not broken anything.

Also:
* have tweaked the map information display to change the text format
to emphasise when it reflects details of a selection (the text turns a bit
orange and is written BOLD) or if the player room (which is what is
referred to if NO rooms are selected) is in this area (the data is written
in BOLD) or is not in the current area shown (the data is written in
ITALICS and not bold).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>