Enhance: make area selection widget respond to use of centerview()
This detail was missing and was confusing when centerview(roomId) was used.

Also, handle the case when the displayed map IS the default area (through
using centerview(...) when the room is in that area AND the area widget has
been set to hide that area name OR it has just been told not to!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>