Revision: route most of map auditing messages to errors.txt file
As the amount of messages produced by the new/improve map auditing/clean-up
code can be large for old map files this commit causes them to be sent to
a profile specific file that is already opened by the Host class but has
not been used for some time.  An option (a check-box) on the "map" tab of
the "profile preferences" dialog controls whether the equivalent
information is also shown on the profile's main console as in the past.

If the option (a global one) which is saved between sessions/runs is NOT
enabled then an advisory is sent to the console advising the user to review
the file contents if a "significant" issue was detected.  If it is enabled
a similar message advising that the information has also been saved. In
either case, as the file is appended to, and not rewritten each time, the
message includes details of the first line of the report so that it can
be located in the file.  One difference between the on-screen and the file
versions of the information is that the former puts up the issues as they
are detected whereas the latter groups the information by subject, first
the general overall issues, then the area ones, in order of the areas' id
numbers and then the room ones, in room order.

Following earlier discussions in these Pull Request I have attempted to
edit all uses of the word "Id" in user visible places to be "id" instead.
This threw up some other messages with issues in several files that I have
tweaked here as well!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>