BugFix: fix crash on profile auto-load (bug 1624720)
The uninitialised TVar::parent member in the default constructor is used
in the profile auto-load (but not the normal one apparently) and was the
cause for segment violations.  In fixing THIS I noted that several other
members of that Class {and worryingly, quite a few *other* classes
throughout the Mudlet application!!!) were not being initialised to
sensible defaults either.  This commit addresses the TVar class, but ONLY
that Class...!

In checking also found that (void) TVar::setNewName( QString ) is not used
{only the two argument (void) TVar::setNewName( QString, int ) that takes
a Lua Type number as a second argument IS used} so commented it out.  Also
found that (int) TVar::nvType and (QString) TVar::nValue members are not
used so removed them as well.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>