C.I.Fix: update to use XCode beta6.2 as 6.1 is "not longer"
Recent error message on Travis MacOS build reports:
"Warning: You are using macOS 10.9.
We (and Apple) do not provide support for this old version.
You may encounter build failures or other breakages.
Please create pull-requests instead of filing issues.
Error: Your Xcode (6.1) is outdated
Please update to Xcode 6.2.
Xcode can be updated from the App Store."

This commit attempt to set the "osx_image" option to be the next oldest one
to try and get C.I. working on that platform - note that according to:

https://docs.travis-ci.com/user/osx-ci-environment#JDK-and-OS-X

the (now broken) default is MacOS 10.9 and XCode beta6.1 - changing to
XCode beta6.2 should not change the OS version - indeed both XCode versions
list seem to come with both MacOS 10.9  and 10.10 - but it is NOT clear how
to select the latter.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>