BugFix: Make TLabel class safe in regard to TEvent attachment
The methods that attach an event (effectively Click, OnEnter and OnLeave)
to a TLabel use a pointer to the relevant TEvent - however that is likely
to be a stack (automatic) allocated instance which will go out of scope
when the caller of these methods end - leaving dangling pointers in the
TLabel.  This commit forces the TLabel to take a deep copy of the pointed
to TEvent's parameters which it uses instead of the originals.