Update: remove Midkemia (and add WoTMUD) to list of predefined MUDs (#321)
* Update: remove Midkemia (and add WoTMUD) to list of predefined MUDs

The I.R.E. MUD Midkemia-online was shutdown on 2016/09/07 and it is not
helpful to continue to offer it as a predefined MUD in the connection
dialog.

This commit removes it from that list and adds a favourite of
mine "WoTMUD" to replace it.  The logo is a temporary one that I will
replace with a similar one that I need to get clearance to use (and release
with a GPL2.1+ licence) from the Wheel of Time MUD sysop with whom I am in
contact {Vivienne needs to get approval from the creator Flash who is away
for a few days...!}

The dlgConnectionProfiles.cpp file has quite a few ASCII strings, as
appropriate for a file that will in the future be subject to I18n work I
took the opportunity to put QStringLiteral(...) wrappers around the
QStrings that are used for non-GUI tasks and tr(...) ones that are.  I
expect it isn't complete but it goes someway towards dealing with THIS
particular file - and it should improve very slightly the generation of
QStrings that need to be constructed because it means as much of possible
of them is done at compile rather than run-time...!

In relation to the above there are points in the code where QStrings are
compared to "" to see if they do not have any contents - the isEmpty()
method is a better way to do the same (it does not involve a string
expression)

In (void)dlgConnectionProfiles::slot_update_name( const QString ) there are
a couple of variables that use a double underscore prefix - that is NOT A
Good Idea: such a prefix is RESERVED in both C and C++. For example see:
[what-are-the-rules-about-using-an-underscore-in-a-c-identifier](http://stackoverflow.com/questions/228783/what-are-the-rules-about-using-an-underscore-in-a-c-identifier)
also a single prefix might be a bit of a problem with libraries...
I have renames the offending things there - but I may have used the single
underscore prefix myself in the past. *blush*

There are some help messages in this class that use an apostrophe in a word
contraction - as a matter of readability this is not considered best
practice IMHO so I have converted them to the full words.

Following experiences in my previous Pull Request I also took a look at
the initialisation and use of some member variables; it became clear that:
* (QString) dlgConnectionProfiles::mOrigin
* (bool) dlgConnectionProfiles::mEditOK
were not used and could be simply deleted and that:
* (QString) dlgConnectionProfiles::mUnsavedProfileName
* (QString) dlgConnectionProfiles::mCurrentProfileEditName
did not need to be member variables and could be local to the method that
used them, so in (void)dlgConnectionProfiles::slot_save_name():
* (QString) dlgConnectionProfiles::mCurrentProfileEditName becomes
  local: (QString) currentProfileEditName
and in (void) dlgConnectionProfiles::slot_addProfile():
* (QString) dlgConnectionProfiles::mUnsavedProfileName becomes
  local: (QString) newName
also added initialisers for:
* (QStringList) dlgConnectionProfiles::mProfileList
* (QPushButton *) dlgConnectionProfiles::connect_button
* (QLineEditQPushButton *) dlgConnectionProfiles::delete_profile_lineedit
* (QPushButton *) dlgConnectionProfiles::delete_button

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* More: two items missed from prior commit

In the previous commit I forgot to take out:
* a check for Midkemia in the dlgProfilePreferences class constructor that
enabled the download map options;
* a similar check in (bool)TMap::restore(QString) that asked if the user
wanted to download a map if there was not one found.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>

* Tweak: Insert MUD supplied Logo/Icon

Received an icon as a .jpg that the MUD operators can allow us to licence
as GPL2.1+ which I have converted to a .png file.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>