bodge: restore minimum font size to profile icon text in Connection Profile
I accidentally restored the font size to a readable one for the text
assigned to the QListWidgetItem for each profile in the Connection Profile
dialog when it was in fact set to a bare minimum to attempt to hide it from
the user's eyes.  This commit undoes that and resets the size to the 1
minimum.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>