Enhance: add sysInterProfileEvent to enable inter-profile signalling
This experimental system allows a script in one profile to:
raiseInterProfileEvent( data ) where data is at least one number, string,
boolean or nil value that can be detected by a declared
sysInterProfileEvent event handler in all active profiles.  The host order
for reception of the event is maintained in a circular manner so that the
profile that emitted the event is the one to receive it last.  The first
argument received after the sysInterProfileEvent one is the name of the
sending host, then the subsequent ones are the arguments supplied to the
invoking function.

Also, as the data to be sent as an event has to be passed up to the main
mudlet class this does have to lock some resources whilst the event is
propagated to the other profiles which means it probably isn't a good idea
for a high volume of traffic.  It, by definition and design, also breaks
the sand-boxing that normally exists between separate profiles.

As the TLuaInterpreter code for this is very similar to the related
raiseEvent(...) function I took the opportunity to "tart-up" that one so
the code is to the same style/standard - that now handles Utf8 encoded
strings correctly and, like the new event also returns the number of
user/script supplied arguments sent in the event.

Whilst coding this function it became clear that there was not a simple
way for a script to establish which Profile it was running in - which is
something that may well be needed when writing scripts for multi-playing
so I have also added a Lua command getProfileName() which takes no
arguments and returns the (possibly Utf-8) string for the profile on which
it is running.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>