Cleanup: clear dead code & prevent reentrancy issue in Host::installPackage
Host::installPackage(...) can call Host::uninstallPackage(...) which in
turn can call Host::installPackage(...) back.  This reentrancy could cause
issues with a single instance of a QDialog used by the former which was
accessed via a class member pointer.  The previous code structure avoids
this, but if the code failed to maintain the case where this is safe (when
second argument has the value of 2) there would be multiple QDialogs
created but only one could be tracked by a class member pointer.  It is
possible to avoid this issue by using a method local pointer which is
enough in the new code which only uses the pointer in this method.

The previous (void)Host::showUnpackingProgress(QString) was rendered
ineffective with an immediate return statement.  As such it was dead code
and removing it also negated the reason for
(void) mudlet::showUnzipProgress(const QString&) so that
(int) TLuaInterpreter::showUnzipProgress(lua_State*) is also dead code.
To avoid any breakage that has been converted into a stub that just reports
via a nil + error message that it is such a stub and can be safely removed
from scripts.

Finally, the QDialog mentioned above was not being removed when
Host::installPackage(...) failed with an error.  This gave the effect of
Mudlet hanging when package/module installation failed.  Note that we are
STILL woefully short of error messages in this process but that needs to
be the subject of further commits.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>