Refactor: remove more dead/ineffective code from Host and dlgTriggerEditor
The (bool)Host::serialize() method was short-circuited with an initial
"return false;" and it was only called in:
(void) dlgTriggerEditor::slot_save_edit() which did not rely on the return
value.  Further examination of the latter method also revealed that all
of the dlgTriggerEditor::slot_saveXXXXAfterEdit() calls it used were just
wrappers around the corresponding dlgTriggerEditor::saveXXX() methods and
there was no other use of the "public slots:" wrappers so they can be
safely eliminated.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>