BugFix: scan for and create needed directories when installing packages
Some file archiving production systems do not seem to be inserting the
(zero length) entries with names ending in a '/' for sub-directories in
archive files.  This commit now makes two passes through any package or
module and identifies all directories that will be needed, it then makes
them before it extracts the files to populate those directories.

This should fix https://bugs.launchpad.net/mudlet/+bug/1413069 as it also
ensures that the proper cleaning process is done should package
installation fail, it removes the "Unpacking" dialog on failure which was
not happening before.

NOTE: There is still an extreme lack of error reporting but as
Host::installPackage(...) is used in several places by different parts of
the application that is probably addressed separately.

Also:
* wrapped strings in Host::installPackage(...) in QStringLiteral(...) or
  tr(...) wrappers and made it more likely to be Utf8 clean.
* Increased the size of the buffer used during extraction of possibly
  compressed archived package files from 100 to 4096 bytes which means
  less loops will be needed to process such archived files.
* Ensured path / file name manipulations/test are not Case Sensitive so
  that they will not misbehave on MacOS platforms.
* Tweaked displaying of the "Unpacking package/module..." dialog so that
  it shows more consistently (it often hid under other windows on my
  GNU/Linux setup!)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>