Refactor: eliminate duplicated set of pointers in dlgTriggerEditor class
Whilst looking at a different issue I spotted that there were two sets of
pointers in the dlgTriggerEditor class that were intended to track the
"current" item in each of Aliases, Buttons ("Actions"), Keys, Scripts,
Timers, Triggers and Variables.  The more appropriately named
mpCurrentXXXXItem ones were largely unused and the mCurrentXXXX ones were
mainly used.  However in the case of the ones for Variables the mixing of
the use of both of them may I though have resulted in some oddities in
hiding and un-hiding of variables that I previously remembered happening.
In fact it doesn't but that is another story...

Also:
* Found some systematic duplicate checks in code for Aliases, Keys etc.
  for a QTreeWidgetItem pointer being non-null which I was able to remove.
  I then spotted a similar test in some Variable handling code in:
  (void) dlgTriggerEditor::slot_var_selected(QTreeWidgetItem *)
  but there were several calls to other methods which used the pointer as
  an argument and it was not immediately clear whether they would modify
  the pointer value or it's pointee.  Further investigation enable me to
  add some const qualifiers to those methods which revealed that the
  address stored in the pointer was not going to be changed so that the
  second test in the method was redundent.

* There was some unneeded C-Style casts in this class and in the cases
  where a cast was needed to resolve signature ambiguities for the compiler
  it is better to use a static_cast in C++ code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>