BugFix: Correct coloration of editor tree icons (part A)
The colours used to represent different folder objects in the editor are
not being coloured consistently, this commit partly addresses this issue
by added a pair of pink folder icons and using them for groups of "Keys"
whose folder was previously cyan in:
(void) dlgTriggerEditor::slot_key_toggle_active()
(void) dlgTriggerEditor::children_icon_key(QTreeWidgetItem *)
(void) dlgTriggerEditor::fillout_form()
but blue in:
(void) dlgTriggerEditor::expand_child_key( TKey *, QTreeWidgetItem *)

Also Alias Groups are violet in all methods in the class except for blue
in:
(void) dlgTriggerEditor::children_icon_triggers(QTreeWidgetItem *)

In correcting this I also added some trivial braces that made the code
layout a bit more regular.

This does not complete the work on sorting out the coloration as some
methods still do not detect the (dummy) folders that are associated with
modules which other modules use brown folder icons for.  That will be in
a following commit.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>