BugFix: Show brown module folders correctly in editor window
Now ensures that module (dummy) folders in the editor window are correctly
coloured when they are activated or deactivated by mouse click or other
means - the previous code that changed between grey and a colour icon
did not detect them and applied a function specific colour instead.

Whilst fixing this I was also able to correct the code that identified
whether a folder (group) container of actions (i.e. buttons) was a top
level one and thus a Toolbar or a sub-folder and thus a Menu, whilst also
detecting toolbars both inside and outside of a module folder.  NOTE: the
code is still not correct as the right hand side of the editor is not
correctly set for Toolbars if they are inside a module.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>